Fixup: Removed TestBookmarkMerging from SyncTests target and moved FailFastTestCase out of Merging source
